requirements.txt
matplotlib==3.10.0
monai==1.4.0
numpy==2.2.1
Pillow==11.0.0
torch==2.5.1
torchvision==0.20.1
tqdm==4.67.1
